+++ README.md
@@ -1,4 +1,4 @@
-# waffle-2022-seminar-react-assignment-1
+# waffle-2022-seminar-react-assignment-2
 
 * cloudfront 배포: https://d6mc9ze3pkchi.cloudfront.net
 * s3 원본: http://joongwon-react-assignment-1.s3-website.ap-northeast-2.amazonaws.com
