+++ src/components/LoginPage/index.tsx
@@ -1,0 +1,3 @@
+export default function LoginPage() {
+  return <div>로그인 페이지</div>;
+}
