+++ src/components/MenuCreatePage/index.tsx
@@ -1,0 +1,3 @@
+export default function MenuCreatePage() {
+  return <div>create menu</div>;
+}
