+++ src/components/MenuDetailsPage/MenuDetails.module.css
