+++ src/components/MenuEditPage/index.tsx
@@ -1,0 +1,3 @@
+export default function MenuEditPage() {
+  return <div>edit menu</div>;
+}
