+++ src/data.json
@@ -21,7 +21,7 @@
     "price": 6000,
     "image": "https://upload.wikimedia.org/wikipedia/commons/1/15/Blueberries.jpg",
     "type": "beverage",
-    "description": ""
+    "description": "이 스무디에는 슬픈 전설이 있습니다"
   },
   {
     "id": 4,
