--- src/pages/StoreListPage.tsx
@@ -1,5 +1,0 @@
-import { Link } from "react-router-dom";
-
-export default function StoreListPage() {
-  return <Link to="/stores/1">세상에서 유일한 와플가게</Link>;
-}
